# frontend/openapi/backend-api.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Formato común y seguro de los errores controlados de la API.",
        "properties": {
          "code": {
            "description": "Código público y estable del error.",
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "correlationId": {
            "description": "Identificador de correlación para soporte y trazabilidad.",
            "example": "7a5f3f75-3bd2-4c47-90fc-6cfc79f1ec2d",
            "type": "string"
          },
          "message": {
            "description": "Mensaje comprensible sin detalles internos.",
            "example": "La solicitud contiene datos inválidos.",
            "type": "string"
          },
          "path": {
            "description": "Ruta de la solicitud que produjo el error.",
            "example": "/api/v1/revocation-requests",
            "type": "string"
          },
          "timestamp": {
            "description": "Fecha y hora UTC del error.",
            "example": "2026-07-20T18:30:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "correlationId",
          "message",
          "path",
          "timestamp"
        ],
        "type": "object"
      },
...
      "DigitalCredential": {
        "properties": {
          "emissionCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "statusListIndex": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "emissionCreatedAt",
          "statusListIndex"
        ],
        "type": "object"
      },
...
      "DigitalCredentialListResponse": {
        "description": "Listado persistido de credenciales digitales de la solicitud autenticada.",
        "properties": {
          "canContinue": {
            "type": "boolean"
          },
          "digitalCredentials": {
            "items": {
              "$ref": "#/components/schemas/DigitalCredentialItem"
            },
            "type": "array"
          },
          "requestStatus": {
            "type": "string"
          }
        },
        "required": [
          "canContinue",
          "digitalCredentials",
          "requestStatus"
        ],
        "type": "object"
      },
...
      "Processing": {
        "properties": {
          "phase": {
            "enum": [
              "SUBMITTING",
              "PROPAGATING",
              "GENERATING"
            ],
            "type": "string"
          },
          "readyAt": {
            "format": "date-time",
            "type": "string"
          },
          "serverTime": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "phase",
          "serverTime",
          "startedAt"
        ],
        "type": "object"
      },
      "Receipt": {
        "properties": {
          "availableAt": {
            "format": "date-time",
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "downloadAvailable": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "PENDING",
              "GENERATING",
              "AVAILABLE",
              "FAILED"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "downloadAvailable",
          "status"
        ],
        "type": "object"
      },
      "RevocationConfirmationRequest": {
        "description": "Decisión completa y consentimiento explícito presentados en el paso 4.",
        "properties": {
          "consentAccepted": {
            "description": "Aceptación expresa del texto mostrado; debe enviarse con valor true.",
            "type": "boolean"
          },
          "consentVersion": {
            "description": "Versión exacta del consentimiento mostrado por el backend.",
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "digitalCredentialUuid": {
            "format": "uuid",
            "minLength": 1,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "type": "string"
          },
          "otherReason": {
            "description": "Descripción requerida únicamente para OTHER.",
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          },
          "reasonCode": {
            "enum": [
              "THEFT",
              "LOSS",
              "DEVICE_OR_NUMBER_CHANGE",
              "SUSPECTED_UNAUTHORIZED_USE",
              "OTHER"
            ],
            "type": "string"
          }
        },
        "required": [
          "consentAccepted",
          "consentVersion",
          "digitalCredentialUuid",
          "reasonCode"
        ],
        "type": "object"
      },
      "RevocationExecutionResponse": {
        "description": "Resultado ciudadano de la revocacion sin exponer identificadores sensibles.",
        "properties": {
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "confirmedAt": {
            "format": "date-time",
            "type": "string"
          },
          "digitalCredential": {
            "$ref": "#/components/schemas/DigitalCredential"
          },
          "firstName": {
            "description": "Primer nombre verificado por ID Perú; ausente solo en evidencia histórica.",
            "example": "ANA",
            "maxLength": 100,
            "type": "string"
          },
          "maskedDni": {
            "example": "******91",
            "type": "string"
          },
          "otherReason": {
            "type": "string"
          },
          "processing": {
            "$ref": "#/components/schemas/Processing"
          },
          "reasonLabel": {
            "type": "string"
          },
          "receipt": {
            "$ref": "#/components/schemas/Receipt"
          },
          "requestStatus": {
            "description": "Estado actual de la solicitud ciudadana de revocación.",
            "enum": [
              "STARTED",
              "CHECKING_AVAILABILITY",
              "NO_DIGITAL_CREDENTIALS_AVAILABLE",
              "PENDING_IDENTITY_VERIFICATION",
              "IDENTITY_VERIFIED",
              "AUTHENTICATED_PENDING_DIGITAL_CREDENTIAL_LIST",
              "CHECKING_DIGITAL_CREDENTIAL_LIST",
              "DIGITAL_CREDENTIALS_AVAILABLE",
              "DIGITAL_CREDENTIALS_SELECTED",
              "REASON_REGISTERED",
              "PENDING_CONFIRMATION",
              "CONFIRMED",
              "REVOCATION_IN_PROGRESS",
              "REVOCATION_SUCCEEDED",
              "REVOCATION_FAILED",
              "REVOCATION_OUTCOME_UNKNOWN",
              "COMPLETED",
              "FAILED",
              "OUTCOME_UNKNOWN",
              "RECEIPT_AVAILABLE",
              "ABANDONED"
            ],
            "type": "string"
          },
          "state": {
            "enum": [
              "PROCESSING",
              "SUCCEEDED",
              "FAILED",
              "OUTCOME_UNKNOWN",
              "RECEIPT_FAILED"
            ],
            "type": "string"
          }
        },
        "required": [
          "digitalCredential",
          "maskedDni",
          "reasonLabel",
          "requestStatus",
          "state"
        ],
        "type": "object"
      },
...
      "RevocationReviewRequest": {
        "description": "Borrador efímero que se valida para presentar el paso 4 sin persistirlo.",
        "properties": {
          "digitalCredentialUuid": {
            "format": "uuid",
            "minLength": 1,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "type": "string"
          },
          "otherReason": {
            "description": "Descripción requerida únicamente para OTHER.",
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          },
          "reasonCode": {
            "enum": [
              "THEFT",
              "LOSS",
              "DEVICE_OR_NUMBER_CHANGE",
              "SUSPECTED_UNAUTHORIZED_USE",
              "OTHER"
            ],
            "type": "string"
          }
        },
        "required": [
          "digitalCredentialUuid",
          "reasonCode"
        ],
        "type": "object"
      },
      "RevocationReviewResponse": {
        "description": "Resumen autoritativo y minimizado de la revocación.",
        "properties": {
          "confirmed": {
            "type": "boolean"
          },
          "confirmedAt": {
            "description": "Fecha UTC persistida; ausente antes de confirmar.",
            "format": "date-time",
            "type": "string"
          },
          "consentText": {
            "type": "string"
          },
          "consentVersion": {
            "maxLength": 64,
            "type": "string"
          },
          "consequences": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "digitalCredential": {
            "$ref": "#/components/schemas/SelectedDigitalCredential"
          },
          "firstName": {
            "description": "Primer nombre verificado por ID Perú; ausente solo en evidencia histórica.",
            "example": "ANA",
            "maxLength": 100,
            "type": "string"
          },
          "maskedDni": {
            "example": "******91",
            "type": "string"
          },
          "otherReason": {
            "description": "Descripción validada para el motivo OTHER.",
            "type": "string"
          },
          "reasonCode": {
            "enum": [
              "THEFT",
              "LOSS",
              "DEVICE_OR_NUMBER_CHANGE",
              "SUSPECTED_UNAUTHORIZED_USE",
              "OTHER"
            ],
            "type": "string"
          },
          "reasonLabel": {
            "type": "string"
          },
          "requestStatus": {
            "description": "Estado actual de la solicitud ciudadana de revocación.",
            "enum": [
              "STARTED",
              "CHECKING_AVAILABILITY",
              "NO_DIGITAL_CREDENTIALS_AVAILABLE",
              "PENDING_IDENTITY_VERIFICATION",
              "IDENTITY_VERIFIED",
              "AUTHENTICATED_PENDING_DIGITAL_CREDENTIAL_LIST",
              "CHECKING_DIGITAL_CREDENTIAL_LIST",
              "DIGITAL_CREDENTIALS_AVAILABLE",
              "DIGITAL_CREDENTIALS_SELECTED",
              "REASON_REGISTERED",
              "PENDING_CONFIRMATION",
              "CONFIRMED",
              "REVOCATION_IN_PROGRESS",
              "REVOCATION_SUCCEEDED",
              "REVOCATION_FAILED",
              "REVOCATION_OUTCOME_UNKNOWN",
              "COMPLETED",
              "FAILED",
              "OUTCOME_UNKNOWN",
              "RECEIPT_AVAILABLE",
              "ABANDONED"
            ],
            "type": "string"
          }
        },
        "required": [
          "confirmed",
          "consentText",
          "consentVersion",
          "consequences",
          "digitalCredential",
          "maskedDni",
          "reasonCode",
          "reasonLabel",
          "requestStatus"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "FlowRefreshCookie": {
        "description": "Refresh JWT rotatorio, transportado exclusivamente mediante cookie HttpOnly.",
        "in": "cookie",
        "name": "revocacion_refresh",
        "type": "apiKey"
      },
      "FlowSessionCookie": {
        "description": "Access JWT de la sesión transaccional, transportado exclusivamente mediante cookie HttpOnly.",
        "in": "cookie",
        "name": "revocacion_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "RENIEC"
    },
    "description": "API institucional para iniciar el flujo ciudadano. La consulta pública exige Google reCAPTCHA v2 Checkbox y la autenticación usa ID Perú con continuidad temporal HttpOnly.",
    "title": "API de revocación de credenciales digitales",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/revocation-requests/current/confirmation": {
      "post": {
        "description": "Guarda la decisión, ejecuta una revocación idempotente y genera la constancia cuando el resultado es exitoso.",
        "operationId": "confirmCurrentRevocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevocationConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationExecutionResponse"
                }
              }
            },
            "description": "Solicitud confirmada o repetición idempotente"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Consentimiento ausente o formato inválido"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Sesión ausente o expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Identidad o paso no permitido"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Versión o decisión incompatible"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Motivo o selección inválida"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Integración de revocación no disponible; la decisión no se persiste"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Persiste la decisión completa y confirma la solicitud",
        "tags": [
          "Confirmación de revocación"
        ]
      }
    },
    "/api/v1/revocation-requests/current/digital-credentials": {
      "get": {
        "description": "Consulta el segundo servicio solo en la primera carga y luego devuelve la instantánea persistida.",
        "operationId": "getCurrentRequestDigitalCredentials",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigitalCredentialListResponse"
                }
              }
            },
            "description": "Listado persistido, incluido el escenario vacío"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Sesión ausente o expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Identidad no verificada o paso no permitido"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Consulta concurrente"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Respuesta externa inválida"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Servicio de listado no disponible"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Timeout del servicio de listado"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Obtiene el listado de credenciales de la solicitud autenticada",
        "tags": [
          "Credenciales vigentes"
        ]
      }
    },
    "/api/v1/revocation-requests/current/execution": {
      "post": {
        "operationId": "resumeCurrentRevocationExecution",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationExecutionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Reanuda idempotentemente una revocación ya confirmada",
        "tags": [
          "Confirmación de revocación"
        ]
      }
    },
    "/api/v1/revocation-requests/current/outcome": {
      "get": {
        "operationId": "getCurrentRevocationOutcome",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationExecutionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Consulta el resultado y la constancia de la solicitud actual",
        "tags": [
          "Confirmación de revocación"
        ]
      }
    },
    "/api/v1/revocation-requests/current/receipt": {
      "get": {
        "operationId": "downloadCurrentRevocationReceipt",
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Descarga la constancia de la sesión autenticada",
        "tags": [
          "Confirmación de revocación"
        ]
      }
    },
    "/api/v1/revocation-requests/current/receipt/retry": {
      "post": {
        "operationId": "retryCurrentRevocationReceipt",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationExecutionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Reintenta únicamente la generación de la constancia",
        "tags": [
          "Confirmación de revocación"
        ]
      }
    },
    "/api/v1/revocation-requests/current/review": {
      "get": {
        "description": "Disponible únicamente después de confirmar; no recupera borradores.",
        "operationId": "getConfirmedRevocationReview",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationReviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Recupera el resumen de una solicitud confirmada",
        "tags": [
          "Confirmación de revocación"
        ]
      },
      "post": {
        "operationId": "previewCurrentRevocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevocationReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationReviewResponse"
                }
              }
            },
            "description": "Resumen vigente sin persistencia del borrador"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Formato inválido"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Sesión ausente o expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Identidad o paso no permitido"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Motivo o credencial inválido"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Valida el borrador y prepara el resumen sin persistirlo",
        "tags": [
          "Confirmación de revocación"
        ]
      }
    },
...
  },
...
}